+++ src/tests/ascent/t_ascent_blueprint_reductions.cpp
@@ -53,7 +53,7 @@
 #include <runtimes/expressions/ascent_blueprint_architect.hpp>
 #include <runtimes/expressions/ascent_execution_manager.hpp>
 #include <runtimes/expressions/ascent_memory_manager.hpp>
-#include <runtimes/expressions/ascent_conduit_reductions.hpp>
+#include <runtimes/expressions/ascent_blueprint_device_reductions.hpp>
 
 #include <iostream>
 #include <math.h>